Add back support for full screen on OS X
The way full screen was handled no longer works so this updates CreateGameWindow to add back support.  On OS X, Apple asks that developers not change the display resolution and instead simply create a window that takes up the entire screen, then modify the back buffer to accommodate the specific drawing size you need if it does not match the display resolution.  According to Apple, they notice when you create a window that takes u the entire screen and optimize your drawing context accordingly.

https://developer.apple.com/library/mac/documentation/GraphicsImaging/Conceptual/OpenGL-MacProgGuide/opengl_fullscreen/opengl_cgl.html
https://developer.apple.com/library/mac/documentation/GraphicsImaging/Conceptual/OpenGL-MacProgGuide/opengl_contexts/opengl_contexts.html#//apple_ref/doc/uid/TP40001987-CH216-SW28